pyproject.toml
[project]
name = "communitymed-ai"
version = "1.0.0"
description = "Multi-Agent TB Screening Assistant powered by MedGemma"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "CommunityMed AI Team"}
]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "trl>=0.8.0",
    "bitsandbytes>=0.43.0",
    "accelerate>=0.30.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "gradio>=4.0.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "loguru>=0.7.0",
    "pyyaml>=6.0.0",
    "kaggle>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]
quantization = [
    "auto-gptq>=0.7.0",
    "optimum>=1.18.0",
]
audio = [
    "librosa>=0.10.0",
    "soundfile>=0.12.0",
]

[project.scripts]
communitymed-train = "scripts.train:main"
communitymed-deploy = "scripts.deploy:main"
communitymed-quantize = "scripts.quantize:main"

[project.urls]
Homepage = "https://github.com/communitymed/communitymed-ai"
Documentation = "https://github.com/communitymed/communitymed-ai#readme"
Repository = "https://github.com/communitymed/communitymed-ai"
Issues = "https://github.com/communitymed/communitymed-ai/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src", "src.agents", "src.models", "src.api", "src.data"]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
